feat: Update code to use big.Int and big.Float for integer and float literals

+++ evaluator/evaluator_test.go
@@ -669,7 +669,7 @@ func TestPowerOperatorStatements(t *testing.T) {
 
 	for _, tt := range tests {
 		evaluated := testEval(tt.input)
-		testFloatObject(t, evaluated, float64(tt.expected))
+		testFloatObject(t, evaluated, tt.expected)
 	}
 }
